--- config.py
@@ -1,6 +1,0 @@
-import os
-
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-
-MODEL_PATH = os.path.join(BASE_DIR, 'models/own.pt')
-DATA_DIR = os.path.join(BASE_DIR, 'FERPlus/data')
